#44 some minor changes
