+++ Code/GeorgiaLibrarySystem-/GTLService/GTLService.csproj
@@ -208,7 +208,6 @@
     <Compile Include="DataAccess\Database\LoginDa_Database.cs" />
     <Compile Include="DataManagement\Code\CopyDm_Code.cs" />
     <Compile Include="DataManagement\Code\LendingDm_Code.cs" />
-    <Compile Include="DataManagement\Code\NoticeDm_Code.cs" />
     <Compile Include="DataManagement\Database\CopyDm_Database.cs" />
     <Compile Include="DataManagement\Database\LendingDm_Database.cs" />
     <Compile Include="DataManagement\IDataManagement\ICopyDm.cs" />
