+++ Code/GeorgiaLibrarySystem-/Tests/UnitTest/LendingTest.cs
@@ -1,4 +1,5 @@
-﻿using Core;
+﻿using System.Collections.Generic;
+using Core;
 using GTLService.DataAccess.Code;
 using GTLService.DataManagement.Code;
 using Moq;
@@ -72,6 +73,34 @@ namespace Tests.UnitTest
 
             //Act
             var result = lendingDm.ReturnBook(copyId);
+
+            //Assert
+            Assert.IsTrue(result == passing);
+        }
+
+        [Test]
+        //pass
+        [TestCase(true)]
+        //fail
+        public void LendingDm_Code_NoticeFilling(bool passing)
+        {
+            //Arrange
+            var mockLendingDa = new Mock<LendingDa_Code>(null);
+            List<Borrow> borrows = new List<Borrow>();
+            mockLendingDa.Setup(x => x.GetAllActiveBorrows())
+                .Returns(borrows);
+            mockLendingDa.Setup(x => x.SaveBorrowChanges())
+                .Returns(true);
+
+            var mockMemberDa = new Mock<MemberDa_Code>(null);
+            Member member = new Member();
+            mockMemberDa.Setup(x => x.GetMember(It.IsAny<int>()))
+                .Returns(member);
+
+            var lendingDm = new LendingDm_Code(mockLendingDa.Object,mockMemberDa.Object);
+
+            //Act
+            var result = lendingDm.NoticeFilling();
 
             //Assert
             Assert.IsTrue(result == passing);
